Module refactoring
