+++ lib/TestCafeTest.js
@@ -1,8 +1,11 @@
 const readDir = require('./utils/rreadDir');
 const fs = require('fs');
+const {exec} = require('child_process');
+const {deleteTempFile, getFilePath, writeTempFile} = require('./utils/localFileHelper');
+
 const tempFilePath = './../src/temp/tempCafe.js';
 const testCafeTestsPath = './../src/testCafeTests/';
-const {exec} = require('child_process');
+
 
 
 class TestCafeTest {
@@ -19,7 +22,7 @@ class TestCafeTest {
         console.log(testFile);
         if(!testFile)
             return;
-        const tempTest = getTestCafeTest(testFile, this.fixtureName, this.Url);
+        const tempTest = composeTestCafeTest(testFile, this.fixtureName, this.Url);
         writeTempFile(tempFilePath, tempTest);
         runTestCafeTest(tempFilePath);
     }
@@ -33,7 +36,7 @@ function runTestCafeTest(tempTestPath) {
 
 }
 
-function getTestCafeTest(path, fixtureName, URL) {
+function composeTestCafeTest(path, fixtureName, URL) {
     let string = '';
     let baseTestCafe = fs.readFileSync(path, {
         encoding: "utf8"
@@ -42,31 +45,5 @@ function getTestCafeTest(path, fixtureName, URL) {
     string += baseTestCafe;
     return string;
 }
-
-
-function deleteTempFile(path) {
-    fs.unlink(path, err => {
-        if(err){
-            console.error(`Failed to delete temp file:
-${err}`);
-        }
-    })
-}
-async function getFilePath(baseDirectory, fileName) {
-    let directory = await readDir(baseDirectory);
-    let d = directory.filter(f => isJsFile(f) && getFileName(f) === fileName);
-    if(!d)
-        console.log(`A file name ${name}.js could not be found
-Make sure you're in the correct directory and try again`);
-    return d.shift();
-}
-
-function writeTempFile(tempFilePath, string){
-    fs.writeFileSync(tempFilePath, string);
-}
-
-const isJsFile = path => path.split('.').pop() === 'js';
-
-const getFileName = path => path.replace(/^.*[\\\/]/, '').split('.').shift();
 
 module.exports = TestCafeTest;
