+++ lib/utils/localFileHelper.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const readDir = require('./utils/rreadDir');
+const readDir = require('./rreadDir');
 
 module.exports = {
